Add Javascript dependencies via yarn
For managing the Javascript dependencies we will use yarn. In this
commit we set up yarn such that packages only needed for development
will go to the tools/ folder, while packages needed for the library
itself will go to the srcjs/ folder.

The development packages will be ignored on GitHub, while the source
code needed by the library will be hosted on GitHub.

Grunt was added as dev dependency and will later be configured to mainly
minify and lint the Javascript and CSS code.

bulma and bulmaswatch have been added in this commit as dependencies
used for the library. This may be later augmented by other libraries.