Add library scaffold
We add all required files to enable R C;D check to succeed. In
particular we provide the necessary meta data in DESCRIPTION.

The first check shows that the inst/ folder will be rather big including
all CSS + map files for all the bulmaswatch themes. Thus, aybe we decide
to skip the map files altogether for a smaller footprint.