Add bulma_page function
`bulma_page` is the entry point for each shiny app using bulma. The
function makes sure the needed resources are added to the webpage. As
bulma itelf is a CSS only framework, this consits basically in loading
the CSS. We added also some bulmaswatch templates to this package, which
can be loaded instead of the original bulma template.

The auto builder tools (grunt) need now to be adapted to re-compile the
bulmaswatch templates, b/c the pre-compiled ones are based onan older
bulma version.