Add SASS compiler to build tools
In this commit we refactored the Gruntfile. Instead of simply copying
the pre-compiled CSS, we compile them afresh using sass. Furthermore, we
decided to not copy the bulmaswatch CSS files to inst/ in order to keep
the package's footprint low. The Gruntfile, however, allows to manually
copy the themes to inst/ in order to facilitate testing.

The overall build strategy is now as follows:
1. Compile bulma & bulmaswatch s[ac]ss files from source
2. Store the minified version in dist/lib/{bulma,bulmaswatch}
3. Copy bulma.min.css to inst/lib/bulma

That is, we need to provide an R function to manually download the
bulmaswatch themes. This strategy helps to keep the file size moderate,
whihc may be important for a future publictaion on CRAN.

TODO: The SASS pre-processor complains about the usage of `/` instead of
`math.div` => create patch file and adapt Gruntfile.js to apply patch to
source code.