Update Gruntfile to install to inst/
We changed the strategy for storing the compiled resources. Earlier we
copied them to dist/ and made yet another copy of needed resources (i.e.
bulma) to inst/, while bulmaswatch was not copied to keep the footprint
low.

With this commit we compile everything (including bulmaswatch) to inst/,
but adapt the .Rbuildignore to not include the css from bulmaswatch. We
want tough to keep the (empty) folder, hence, we need to add a flag to
the DESCRIPTION file.

This setuo has the advantage that the templates are available while
locally developping, but do not appear in the installed version. For the
installed version, the download function (yet to be developped) will be
needed to install the templates post installation.