Add R Script to replace google fonts to local copy
We added a R script which reads in all passed CSS files and looks for
the @import directive. The inforation from there is then parsed and
gfonts::download_font is used to get a local copy of all fonts, such
that the font can be used with almost any browser.

The @import directive is then replaced by a generic CSS rule which
allows including the font in most modern browsers.

To automatize this task, grunt-run is used to initiate this script.
Unlike other plugins, it does not know a `src` argument. Hence, a bit of
adaptations are needed (especially how to reuse existing configs). The R
script was eventually changed to accept one comma separated string of
file names, which it processes one after the other. This hack is
basically needed, b/c we did not manage to pass a proper dynmic string
with file names from Grunt to R.

TThe color detection does not work properly, hence, we set the color
option explicitely in the R script.