Add data storing colors and font used for bulma
Ww created a script which parses both, the official bulma documentation
and the SASS files to extract all color and font declarations and
exposes them to an R data frame. The idea of this script is to allow R
users to conventiently use the same colors than the bulma framework.

We parede the docuemtation to get a list of all defined colors.
Theoretically, we could also retrieve the values from there, but decide
to parse them from the SAA files directly.

We do not store colors which are the resultant of a function, but only
the ard coded values, respectively the ones which are aliases of hard
coded values.

In a next step we will do the same for the themes, where we will source
this script to have the defualts as not all varaibels will be re
overridden by the templates.