Change strategy to compile SASS instead of parsing
Instead of parsing the values from the respective SASS files, we decided
to create a temporary SCSS files which imports the relevant SASS files
and _compile_ it directly. In this way we can also get the values for
derived variables.

We do that by first checking the documentation under
<https://bulma.io/documentation/customize/variables/> to get all color
variable names (derived, compound  or literal). Then we create a
temporary SCSS file where we create a rule for a class named exactly
like the color (e.g. `.black`) and import the necessary bulma SASS files
(initial-variables, derived-variables and functions). Finally, we
compile it using the Dart compiler sitting in folder tools/.

We needed to run directly via the command line approach, as the R
package SASS uses an older compiler which would lead to inconsistencies
with the files processed to be usable by the Gruntfile cinfiguration.