Add a generic function to extract colors
To get the color codes (and Google font used) we exported the code form
`bulma_config` to more generic format, where we can also include the
bulmaswatch SCSS in order to not replicate code.

Additionally, the datasets are properly documented. Eventhough we do not
provide the bulmaswatch themes by default, but will provide a function
to install them post-install, we decided to keep the data in the
package, which is somehow an inconsistency.