Add minified CSS and font files
The general approach will be that bulmaswatch themes will __not__ be
installed by default, but the library will provide a function to
download the respective themes and fonts post-install form the github
repo.

In this commit, we add all files, in order to be able to test the
functionality of the download script consequently.