Fix font assignment in data generating scripts
A wrong regex added too many characters to the font leading to an
erroneous assignment in `bulmaswatch_config`. This commit fixes this.
Additionally, the data is now stored as `data.frame` instead of a
`tibble` to keep dependencies of the library low. The data includes now
an `id` field for fonts, to be able to use that for the download script.
The `name` is still part of the data, but is now stored in `name` rather
than `Google Font`.