Add Bulma Theme Downloader
In this commit we add a function to download the pre-compiled themes
from the GitHub repository. This function uses the GitHub REST API to
fect all releveant download links.

Additionally, the .Rinstignore file was added, to avoid that
devtools::build add files in the inst/lib/bulmaswatch and inst/fonts
folder while building.

Finally, the markdown extension for roxygen2 was switched on and new
roxygen doc strings can now benefit from markdown sytnax. Most notably,
the possibility to run R code (which is done for the function
`download_bulma_themes`.