Adapt .Rprofile to properly close terminal
In the previous version the terminal was not properly closed resulting
in a hung up R session whenever we restarted the session. By switching
form `rstudioapi::terminalExecute` to `rstudioapi::terminalSend` we
managed to close the terminal properly.

We needed to add "\r\n" to execute the function which we sent. To run
the cleanup code we used `reg.finalizer` on the .GlobalEnv, which runs
code whenever the environment provided (here .GlobalEnv) is garbage
collected (hwich apparently happens upon restart).

An earlier attempt to use `.Last` did not work.