Change data representation of bulma{swatch}_config
In the previous version, data was explicitely exported
(`usethis::use_data(...internal = FALSE`). This made it neessary, that
accessing data from _own_ functions worked only via accessing it via the
namespace operator (i.e. `shinyBulma2::bulma_config`).

After reading <https://tinyurl.com/22k7tknd>, we decided to change the
strategy. The data is set up for internal usage, but by exporting it
explicitely, we make it also available for outside use. In this way we
get bets of both worlds: we do not need to namesapce the data, yet it is
also exported to users of the library.