Fix .Rprofile to run exit code only conditionally
In the previous version, the exit code in .Rprofile ran regardless of
whether Rstudio was running or not. While this error does not occur
while indeed working in Rstudio, it appears when running build
(apparently this is ran in an own instance).

Thus, we added a condition, to run the exit code only if Rstudio is
indeed running.