Add testthat unit testing framework
In order to allow for unit tests, we added the `testthat` framework to
the project. A simple test that no themes are present in a factory fresh
install was added.

Some thoughts need to go into that, as for local development themes are
installed ba design and it would be ncie to chekc that too.