Fix font path in  CSS
In order to use the fonts, the font path must be properly set. We will
use `addResourcePath`, whcih maps a physical directory to a prefix (e.g.
"fonts"). Then, we can address resources there via "/fonts" (not the
leading slash. Thus, the resource path prefix and the path given in the
css for the fonts must be the same, suc that the CSS files can find the
fonts.