Add size/offset validation routine for columns
In this commit, we add a funciton (+unit tests) to check whether a given
size/offset is valid and if so, transforms that to the proper bulma
class.

This funciotn is heavily influenced by htmltools::validateCssUnit. Since
offsets and sizes share almost the same options, we decided to write an
internal workhorse function. This has the nasty side effect that
validation errors are raised under the name of the helper function.