Patch superseded function call html_node
In the data creation script the function `html_node` was used. After
consulting the docs, it turned out that this function is superseded by
`html_element`. This commit replaces the functions to the newer version.
(N.B. the reason why the funciton was there in the first place was
probably some copy & paste from an internet resource).