Adapt unit tests for bulma columns layout
Due to the change of the API of `bulma_columns` and `bulma_column`, we
needed to adapt the unit tests for those functions. The additional
opportunity to specify gaps is now alos included in the test suite.
Tests for warnings were added, as well as the possibility to test
multiple sizes/offsets/gaps.