Add column to configurtaion datasets
For certain bulma colors, we can use the (simplified) class `is-{color}`
for example `is-primary`. The colors for which those helpers are
defined, are the names of the SASS map `$colors`. In this commit, we
changed the data generating script to inlcude this information as column
`is_color_map_key` to the data, to allow for using it as a validation
set in elements which use such color shortcut classes.