Add an unexported environment to store globals
We introduce an unexported environment (with the empty environment as
parent), where we want to store globals. Most notably the theme used.
This helps in subsequent validation calls to test versus the actual
theme used. Not all theme for instance define the same color variables.
A validate_* function which wants to validate the colors, needs thus to
know which theme we are using to chekc the respective lines in
`bulmaswatch_config`.