Add color_class column to bulma{swatch}_config
In order to test whether a color can be used via
`has-(text|background)-.*`, we added a column to the config data frames
holding a comma separated string with tlements `text` or `background`
respectively to indicate whether this particular color can be used.

The information is scrpaed from the official bulma documentation.

This feature is useful to properly implement a color validator function.