Add validate function for bulma colors
To check bulma colors we make use of the `.*_config` datasets, which
contain a column `color_class`, which indicates whether this particular
color can be used via `has_(text|background)_{color}`. In future
versions we may want to add the possibility to use `is-{color}`
specifiers as well.

TODO: add unit tests.