Add unit tests for validate_bulma_color
Some minor changes were done in `utils.R` to allow for not collapsing
several classes. Collapsing makes sense for elements where we want to
pass several widths (for instance) for a _single_ element. For `color`
though we still may want to pass several colors (to test them in one
go), but we will liely not assign them to the same element. hence, we
need the un-collapsed result form the helper function.

We also added some additional information to the color validation
routine, when we are explicitely asking to test only primary colors, but
pass a non-primary color. This distinction should help the end user to
understand that the color is indeed valid, but is not a key color and if
requested ot validate only against those, the validation is expected to
fail. The addiitonal information makes it just easier to udneratnd this
fact.