Adapt unit tests and implement tag expectations
To test HTNL tags several helper expectations were added, which make
sure that the tags are as expected. For instance we can test whether the
number of child elements matches the expectation or the type and class
of the child elements are corresponding to the expectation.

This framework helps to test tags in a more structured way rather than
comparing the string represntation of the generated tags.

TODO: review existing unit tests and replace string comparison with the
new framework.