Add button color class
The column `color_class` describes for which occasions the color can be
used. The particular case of `ghost` made it necessary to add a class
`button` to indicate that a certain color can be used only for buttons.
For buttons "normal" background colors can be used as well, thus, we
needed to add  *button*  to these color classes as well.