Add bulma notification
The bulma notification elment is styled with a `delete` button. This,
button is currently just an HTML element and has no functionality yet.
Later, we will develope Shiny bound inputs, where we will define a
`shiny_action_button` on top of `shiny_button`, The, we may want to
revisit `bulma_notification` to replace th delete HTML by an input
elmenet, which implements the removing functionality in JS.