Add validate_bulma_size function
As size classes are used at many differnt places throughout the bulma
framework, we observed that we duplicated the validity checker for
proper sizes quite often. Thus, we decided to move this check into a
function of its own.

As we originally named the validation check function for bulma columns
`validate_bulma_size`, we had to rename this original function to
`validate_bulma_column_size?  which anyways reflects better what the
inten of this function is.