Add Bulma Table
In this commit we added the function to create a bulma table from a
matrix like object. The function is basically a copy and paste from
`shiny::renderTabler`. The latter's workhorse is `xtable::xtable`,
which, however, returns the HTML in a string (and not in a shiny tag).
Thus, we added a helper function which parses this string back to a
shiny tag. This required to add another dependency (in `Suggests`
though) `{xml2}` for the parsing. We make this parsing optional, thus we
could avoid adding the dependency as `Imports`.