Add Unit Tests for bulma_table
Tests were added for `bulma_table`. As with the other bulma elments,
most of the tests rely on the `expect_tag_*` family of expectations. As
`bulma_table` is special to the extent that the workhorse function
returns an HTML string (and not tags), we need to rely on the `{xml2}`
library to first transform the HTML strings to tags. This works,
however, only if it is available on the system. Thus, we needed to skip
these test suit on systems where `{xml2}` is not available.