Add Unit Tests for Bulma Tag
The final test suite for the bulma elements section has been completed
with this commit. As usual, we use the homebreed expectations to check
the structure of the generated HTML and check aditionally for warnings /
errors generated by the basic argument checking.

We followed the idea that wranings / errors generated by internal
validation funcitons are not (again) tested, as these tests happen
anyways in the respective uit tests for these validation tests. This
saves duplication and should be the very idea of **unit** testing.