Update bulma_notification and some examples
While checking the code for `bulma_columns` (to draw somw inspiration
for `bulma_container regarding handling of media breakpoints), we
observed that sone code in the examples constructed bulma elements by
hand, rather than relying on the developped function. From the
development history that makes perfectly sense, because at the time of
writing an example for `bulma_columns`, say, other elements were not yet
implemented.

Of course this has to be changed, whenever a new element is introduced.
This replaement steps have to happen every now and then to make sure no
hand crafted elements remain, where there is a proper lement in this
package.

In doing this replacmeent, we figured that it would be nice, if we also
could create a `bulma_notification` without the delete button, so we
implemented a new parameter.

In this cleaning step, we also touced `bulma_page`. However, there is
still some hand crafted HTML code, which eventually needs replacement,
once the functions are implmented.