Add bulma_level
To add this element, we also needed to refactor some of the old code, as
we wanted to add a `bulma_heading` function (which adds a `heading`
class to an element). This name, however, was already taken by the
workhorse function for `bulma_title`. We renamed this function to
`bulma_header` and introduced `bulma_heading` to be used in the bulma
level element.

The bulma level element needs to add items either in the left or right
part of the level. For this we need an interface which allows ot add
several elements to either of these sides. A `...` aproach would not
works, as we need two places to add a variable amount of elements. Thus,
we decided to ask for a single argument (`left`, `right`) which expects
a listish element. The function then loops through all celement in this
list (typically a `tagList`) and wraps the elements (if necessary) in
`bulma_level_item`.

The latter can also be used to create level items directly. Either by
specifiying heading and title (which creates KPI like visuals), or by
`content`, which allows for full flexibility, becuase the elements are
taken as provided.