Update expect_tag_class
We updated expect_tag_class in a way that unmatched classes in the
reference can raise a no-match. This makes it possible to test for the
absence of a class. We provide all classes which should be there, and if
there are some in the reference, setting `exact = TRUE` will raise a
failure.