Add bulma Media Object
The media object is a rather easy element with a left, a right and a
content part. While impemeting the examples from the web site, we
figured that sometimes we want to wrap an image not a
`<figure>`environment but in another, `p`, say.

Thus, we decided to add a parameter to `bulma_iamge.