Finalize unit tests for bulma_hero
We finalized the unit tests for bulma_hero. This required to make some
if clauses more verbose b/c the previous approach

```
div(if (!is.null(x)) x)
```

lead to some `NULL` elements in the output. Thus, we replaced this
pattern by a classical if clause.

Furthermore, we added unit tests for some of the helpers for the
validate routines.